requirements.txt
alembic==1.4.2
altgraph==0.17
appdirs==1.4.4
asgiref==2.3.2
astroid==2.4.1
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.0
channels==2.1.0
chardet==3.0.4
click==7.1.2
cmake==3.17.3
colorama==0.4.3
constantly==15.1.0
contextvars==2.4
convertdate==2.2.1
cryptography==2.9.2
cycler==0.10.0
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
dlib @ https://files.pythonhosted.org/packages/0e/ce/f8a3cff33ac03a8219768f0694c5d703c8e037e6aba2e865f9bae22ed63c/dlib-19.8.1-cp36-cp36m-win_amd64.whl#sha256=794994fa2c54e7776659fddb148363a5556468a6d5d46be8dad311722d54bfcf
entrypoints==0.3
et-xmlfile==1.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==1.1.2
flask-marshmallow==0.9.0
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.1
future==0.18.2
holidays==0.10.2
hyperlink==19.0.0
idna==2.9
image==1.5.32
immutables==0.14
importlib-metadata==1.6.1
importlib-resources==2.0.1
incremental==17.5.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==2.13.6
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mysql-connector-python==8.0.20
nb2xls==0.1.6
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.5
opencv-python==4.2.0.34
openpyxl==3.0.3
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.0
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.1.2
pip-autoremove==0.9.1
pipenv==2020.6.2
prometheus-client==0.8.0
prompt-toolkit==3.0.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==1.9.6
Pygments==2.6.1
PyHamcrest==2.0.2
PyInstaller==3.6
pylint==2.5.2
PyMeeus==0.3.7
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypng==0.0.20
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
SQLAlchemy==1.2.13
sqlparse==0.3.1
terminado==0.8.3
terminal==0.4.0
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typing==3.7.4.1
urllib3==1.25.9
virtualenv==20.0.23
virtualenv-clone==0.5.4
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
wrapt==1.12.1
XlsxWriter==1.2.9
zipp==3.1.0
zope.interface==5.1.0
